+++ sdk/nodejs/cloudwatch/index.ts
@@ -9,6 +9,7 @@ export * from "./eventPermission";
 export * from "./eventRule";
 export * from "./eventRuleMixins";
 export * from "./eventTarget";
+export * from "./flowWidgets";
 export * from "./getLogGroup";
 export * from "./logDestination";
 export * from "./logDestinationPolicy";
@@ -21,4 +22,6 @@ export * from "./logSubscriptionFilter";
 export * from "./metric";
 export * from "./metricAlarm";
 export * from "./metrics";
+export * from "./simpleWidgets";
+export * from "./widgetJson";
 export * from "./widgets";
