+++ sdk/nodejs/tsconfig.json
@@ -129,6 +129,7 @@
         "cloudwatch/eventRule.ts",
         "cloudwatch/eventRuleMixins.ts",
         "cloudwatch/eventTarget.ts",
+        "cloudwatch/flowWidgets.ts",
         "cloudwatch/getLogGroup.ts",
         "cloudwatch/index.ts",
         "cloudwatch/logDestination.ts",
@@ -142,6 +143,8 @@
         "cloudwatch/metric.ts",
         "cloudwatch/metricAlarm.ts",
         "cloudwatch/metrics.ts",
+        "cloudwatch/simpleWidgets.ts",
+        "cloudwatch/widgetJson.ts",
         "cloudwatch/widgets.ts",
         "codebuild/index.ts",
         "codebuild/metrics.ts",
